Some manny changes x2

+++ app/Http/Controllers/Admin/ProductController.php
@@ -7,44 +7,47 @@ use App\Http\Requests\ProductRequest;
 use App\Models\MoneyType;
 use App\Models\Product;
 use App\Models\ProductType;
-use Illuminate\Http\Request;
+use App\Models\TourType;
+use App\Traits\ResponseTrait;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
+use Illuminate\View\View;
 
 class ProductController extends Controller
 {
+    use ResponseTrait;
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Contracts\View\View
+     * @return View
      */
-    public function index()
+    public function index():View
     {
-        $products = Product::all();
-
-        return view('admin.products.index')->with('products',$products);
+        return view('admin.products.index')
+            ->with('products',Product::all());
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Contracts\View\View
+     * @return View
      */
-    public function create()
+    public function create():View
     {
-        $moneyTypes = MoneyType::all();
-
-        $types = ProductType::all();
-
-        return view('admin.products.create')->with('moneyTypes',$moneyTypes)->with('types',$types);
+        return view('admin.products.create')
+            ->with('moneyTypes', MoneyType::all() )
+            ->with('types',ProductType::all() )
+            ->with('tourTypes',TourType::all() );
     }
 
     /**
      *  Store a newly created resource in storage.
      *
      * @param ProductRequest $request
-     * @return \Illuminate\Http\RedirectResponse
+     * @return RedirectResponse
      */
-    public function store(ProductRequest $request)
+    public function store(ProductRequest $request):RedirectResponse
     {
         DB::beginTransaction();
         try{
@@ -54,45 +57,52 @@ class ProductController extends Controller
                 'price' => $request->request->getInt('price') ,
                 'type' => $request->request->getInt('type'),
                 'money' => $request->request->getInt('money'),
+                'tourType' => $request->request->getInt('tourType') == 0
+                    ? null
+                    : $request->request->getInt('tourType'),
             ]);
 
             DB::commit();
         }catch (\Exception $e){
+            DB::rollback();
 
-            app()->hasDebugModeEnabled() ? $message = $e->getMessage() : $message = __('app.error_create', ['object' => __('app.tour_states_singular')]) ;
-
-            return redirect()->route('product.create')->with('message',$message);
+            return $this->errorResponse('product.create' , $e->getMessage(), __('app.error_create', ['object' => __('app.products_singular') ]) );
         }
-
-        return redirect()->route('product.index')->with('success', __('app.success_create ',['object' => __('app.products_singular')] ));
+        return $this->successCreateResponse('product.index',__('app.products_singular'));
     }
 
     /**
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Contracts\View\View
+     * @return View
      */
-    public function show($id)
+    public function show(int $id):View
     {
-        $product = Product::findOrFail($id);
-
-        return view('admin.products.show')->with('product',$product);
+        return view('admin.products.show')
+            ->with('product', Product::findOrFail($id) );
     }
 
     /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Contracts\View\View
+     * @return View
      */
-    public function edit($id)
+    public function edit(int $id):View
     {
         $product = Product::findOrFail($id);
-        $moneyTypes = MoneyType::all();
-        $types = ProductType::all();
 
-        return view('admin.products.edit')->with('product',$product)->with('moneyTypes',$moneyTypes)->with('types',$types);
+        if($product->price == null)
+        {
+            $this->errorAbort404();
+        }
+
+        return view('admin.products.edit')
+            ->with('product',$product )
+            ->with('moneyTypes', MoneyType::all() )
+            ->with('types', ProductType::all() )
+            ->with('tourTypes', TourType::all());
     }
 
     /**
@@ -100,54 +110,60 @@ class ProductController extends Controller
      *
      * @param  ProductRequest  $request
      * @param  int  $id
-     * @return \Illuminate\Http\RedirectResponse
+     * @return RedirectResponse
      */
-    public function update(ProductRequest $request, $id)
+    public function update(ProductRequest $request,int $id):RedirectResponse
     {
         DB::beginTransaction();
         try{
+
             Product::whereId($id)->update([
                 'name' =>$request->request->get('name'),
                 'description' => $request->request->get('description'),
                 'price' => $request->request->getInt('price') ,
                 'type' => $request->request->getInt('type'),
                 'money' => $request->request->getInt('money'),
+                'tourType' => $request->request->getInt('type') != 1
+                    ? null
+                    : $request->request->getInt('tourType')
             ]);
 
             DB::commit();
         }catch (\Exception $e){
             DB::rollback();
 
-            app()->hasDebugModeEnabled() ? $message = $e->getMessage() : $message = __('app.error_update', ['object' => __('app.products_singular') ]) ;
-
-            return redirect()->route('product.edit')->with('message',$message);
+            return $this->errorResponse('product.edit' , $e->getMessage(), __('app.error_update', ['object' => __('app.products_singular') ]) );
         }
-        return redirect()->route('product.index')->with('success' ,__('app.success_update ',['object' => __('app.products_singular') ]));
+        return $this->successUpdateResponse('product.index', __('app.products_singular') );
     }
-
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
-    public function destroy($id)
+    public function destroy(int $id):Response
     {
         DB::beginTransaction();
         try{
             $product= Product::findOrFail($id);
+
+            if($product->price == null)
+            {
+                return $this->errorDestroyResponse( new \Exception(), __('app.error_delete'), 500 );
+            }
+
             $product->delete();
 
             DB::commit();
         }catch (\Exception $e){
             DB::rollback();
 
-            app()->hasDebugModeEnabled() ? $message = $e->getMessage() : $message = __('app.error_delete');
-
-            return response($message,500);
+            return $this->errorDestroyResponse( $e , __('app.error_delete'), 500 );
         }
-        return response( __('app.success'),200);
+        return $this->successDestroyResponse(__('app.success'));
     }
+
 
 }
